Connect Four game
Connect Four is a game in which the players choose a color and then take turns dropping colored tokens into a seven-row, seven-column vertically suspended grid. The pieces fall straight down, occupying the lowest available space within the column. The objective of the game is to be the first to form a horizontal, vertical, or diagonal line of four of one's own tokens. I have created this using html,css and javascript.